requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
requests==2.25.1
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.3
vk-api==11.9.1
python-dotenv~=0.20.0

alembic~=1.7.7